Removing the useless segmentation of the UI, just made it a pain. THIS BUILD IS NON FUNCTIONAL DO NOT INSTALL
